Update apihelper.html
With the file data.js in the same directory, your login fields are prefilled with the data provided in data.js.   
 
Functionality: Attemt to post a request to a web-app (google script) with apihelper.html run from local harddrive ends up in a CORS failure. This version of apihelper.html starts to deal with this problem (still not working).